+++ include/tulips/transport/Producer.h
@@ -34,7 +34,7 @@ public:
    *
    * @return the status of the operation.
    */
-  virtual Status commit(const uint32_t len, uint8_t* const buf,
+  virtual Status commit(const uint16_t len, uint8_t* const buf,
                         const uint16_t mss = 0) = 0;
 
   /**
