+++ src/transport/shm/Device.cpp
@@ -128,7 +128,7 @@ Device::prepare(uint8_t*& buf)
 }
 
 Status
-Device::commit(const uint32_t len, uint8_t* const buf,
+Device::commit(const uint16_t len, uint8_t* const buf,
                UNUSED const uint16_t mss)
 {
   m_log.trace("SHM", "committing packet: ", len, "B, ", (void*)buf);
